fix: Fix repeated file uploads caused by BOM
Because the excerpted files in the webdav tracking directory are modified
directly in the server (delete the BOM mark), the excerpted files on
the webdav server and the terminal device are judged to be different, and
the file content is repeatedly uploaded to server. The current solution is to
copy all the files in the "摘抄" directory tracked by webdav to another
directory (formatted-excerpt). Modifications in this directory will not
cause modifications to the webdav tracking directory, thus ensuring
communication with the terminal device consistency.